+++ SPC_Coopenae/SPC_Coopenae.UI/Areas/Mantenimientos/Models/Productos.cs
@@ -9,7 +9,7 @@ namespace SPC_Coopenae.UI.Areas.Mantenimientos.Models
     {
         public int IdProducto { get; set; }
         public string Nombre { get; set; }
-        public double ComisionGenera { get; set; }
+        public decimal ComisionGenera { get; set; }
         public int EsDolar { get; set; }
         public int Estado { get; set; }
